+++ src/lib.rs
@@ -1,4 +1,8 @@
 pub mod matrix;
+pub mod vector;
 
 #[allow(unused_imports)]
 pub use matrix::*;
+
+#[allow(unused_imports)]
+pub use vector::*;